adding base test class for unit tests in routes
